Factored out record value validation
