+++ netbox_dns/validators/__init__.py
@@ -1,2 +1,3 @@
 from .dns_name import *
+from .dns_value import *
 from .rfc2317 import *
